+++ lib/bundler/install_dash_docs.rb
@@ -1,6 +1,7 @@
 # frozen_string_literal: true
 
 require_relative "install_dash_docs/version"
+require_relative "install_dash_docs/bundler_plugin"
 require_relative "install_dash_docs/cli"
 require_relative "install_dash_docs/dash"
 require_relative "install_dash_docs/dash_install_url"