+++ plugins.rb
@@ -1,1 +1,2 @@
+# This file is loaded by bundler plugin subsystem, simply requiring gem's code is enough
 require "bundler/install_dash_docs"